+++ src/components/tabScreens/FinalMeasurementDetails.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { useBookingLeadById } from "@/hooks/booking-stage/use-booking";
 import { useFinalMeasurementLeadById } from "@/hooks/final-measurement/use-final-measurement";
 import { useAppSelector } from "@/redux/store";
 import { motion } from "framer-motion";
