+++ src/components/tabScreens/OpenLeadDetails.tsx
@@ -14,9 +14,6 @@ import {
 import { formatDateTime } from "../utils/privileges";
 import { useLeadById } from "@/hooks/useLeadsQueries";
 import { useAppSelector } from "@/redux/store";
-import { useState } from "react";
-import { Button } from "../ui/button";
-import AssignLeadModal from "../sales-executive/Lead/assign-lead-moda";
 
 type OpenLeadDetailsProps = {
   leadId: number;
